Fix green smearing in h264 decoding Occasionally when playing h264-encoded files, green blocks will appear at the edges of the frame, and get smeared around (following the motion vectors, of course).  This bug has existing in Xine for well over a year, and I keep hoping it will be fixed, but it never was.
I recently learned that gstreamer had a similar problem and it was fixed
last year.  With this as a hint, I found
http://bugzilla.gnome.org/show_bug.cgi?id=364139 and the patch attached
to that bug.

I've adapted that patch to xine, and it is attached.  I have compared
clips before and after applying this patch, and in all cases the green
artifacts have gone away, and I have seen no negative side effects.